sync: uses contextvar to store executor reference
The asgiref "Local" utility does not maintain its value when the
client code (inside a async_to_sync wrapped function for example)
launches a task using e.g. asyncio.create_task. This means that
the "AsyncToSync.executors.current" reference is lost and causes
a deadlock.

Adds a test for this deadlock condition and uses the contextvars
module instead of asgiref.local to store the reference to the
current thread executor.